Extract controller event loop
Move the dispatch loop and event handling out of main so the
CLI entrypoint stays focused on process wiring.

Keep sysfs polling as an external source and let the
controller own internal event translation and handling.
